editoast: move BufferStopCache to infra_cache
