+++ editoast/src/infra_cache/object_cache.rs
@@ -1,0 +1,3 @@
+mod buffer_stop_cache;
+
+pub use buffer_stop_cache::BufferStopCache;
