+++ editoast/src/infra_cache/object_cache/buffer_stop_cache.rs
@@ -1,0 +1,60 @@
+use crate::infra_cache::Cache;
+use crate::infra_cache::ObjectCache;
+use crate::schema::BufferStop;
+use derivative::Derivative;
+use diesel::sql_types::Double;
+use diesel::sql_types::Text;
+use editoast_schemas::primitives::OSRDIdentified;
+use editoast_schemas::primitives::OSRDTyped;
+use editoast_schemas::primitives::ObjectType;
+
+#[derive(QueryableByName, Debug, Clone, Derivative)]
+#[derivative(Hash, PartialEq)]
+pub struct BufferStopCache {
+    #[diesel(sql_type = Text)]
+    pub obj_id: String,
+    #[derivative(Hash = "ignore", PartialEq = "ignore")]
+    #[diesel(sql_type = Text)]
+    pub track: String,
+    #[derivative(Hash = "ignore", PartialEq = "ignore")]
+    #[diesel(sql_type = Double)]
+    pub position: f64,
+}
+
+impl OSRDTyped for BufferStopCache {
+    fn get_type() -> ObjectType {
+        ObjectType::BufferStop
+    }
+}
+
+impl OSRDIdentified for BufferStopCache {
+    fn get_id(&self) -> &String {
+        &self.obj_id
+    }
+}
+
+impl Cache for BufferStopCache {
+    fn get_track_referenced_id(&self) -> Vec<&String> {
+        vec![&self.track]
+    }
+
+    fn get_object_cache(&self) -> ObjectCache {
+        ObjectCache::BufferStop(self.clone())
+    }
+}
+
+impl BufferStopCache {
+    pub fn new(obj_id: String, track: String, position: f64) -> Self {
+        Self {
+            obj_id,
+            track,
+            position,
+        }
+    }
+}
+
+impl From<BufferStop> for BufferStopCache {
+    fn from(stop: BufferStop) -> Self {
+        Self::new(stop.id.0, stop.track.0, stop.position)
+    }
+}
