+++ editoast/src/views/infra/auto_fixes/buffer_stop.rs
@@ -4,7 +4,7 @@ use tracing::debug;
 
 use super::new_ref_fix_delete_pair;
 use super::Fix;
-use crate::schema::BufferStopCache;
+use crate::infra_cache::object_cache::BufferStopCache;
 use crate::schema::InfraError;
 use crate::schema::InfraErrorType;
 use editoast_schemas::primitives::OSRDObject as _;
